% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{averagePerBin}
\alias{averagePerBin}
\title{Find the average values of x in provided bins}
\usage{
averagePerBin(x, bins)
}
\arguments{
\item{x}{GRanges with values to find average of}

\item{bins}{GRanges with bins to average over}
}
\value{
matrix
}
\description{
Uses a method similar to the GenomicRanges tutorial for finding binned averages. Note, this method may not return the
same results as averagePerBinBSseqLike().
}
\examples{

gr1 <- GRanges(seqnames = "chr2", ranges = IRanges(103, 106), strand = "+", score = 5L, GC = 0.45)
gr2 <- GRanges(
    seqnames = c("chr1", "chr2"),
    ranges = IRanges(c(107, 113), width = 3),
    strand = c("+", "-"),
    score = 3:4, GC = c(0.3, 0.5))
gr3 <- mergeMCols(gr1, gr2)

bins <- GRanges(seqnames=c("chr1","chr2"), ranges=IRanges(c(100, 100), width=15))

averagePerBin(gr3, bins)

}
