% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEpibed.R
\name{epistateCaller}
\alias{epistateCaller}
\title{Calculate and plot a hierarchical clustering of the epibeds}
\usage{
epistateCaller(
  mat,
  stringdist_method = "hamming",
  hclust_method = "ward.D2",
  plot = TRUE
)
}
\arguments{
\item{mat}{Input matrix that comes out of tabulateEpibed()}

\item{stringdist_method}{stringdist::stringdist algorithm to use (default: "hamming")}

\item{hclust_method}{Clustering algorithm to use (default: "ward.D2")}

\item{plot}{Whether to plot the clustered epibeds (default: TRUE)}
}
\value{
A hierarchical cluster (hclust) object
}
\description{
WARNING: This is a beta function and should not be used for publication level analyses!!!!
}
\examples{

#epibed.nome <- system.file("extdata", "hct116.nome.epibed.gz",
#                           package="biscuiteer")
#epibed.nome.gr <- readEpibed(epibed = epibed.nome, is.nome = TRUE,
#                             genome = "hg19", chr = "chr1")
#epibed.tab.nome <- tabulateEpibed(epibed.nome.gr)
#epistateCaller(epibed.tab.nome)

}
