% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSeqLengths}
\alias{getSeqLengths}
\title{Get the seqlengths of a chromosome or all chromosomes}
\usage{
getSeqLengths(gen = c("hg38", "hg19", "mm9", "mm10"), chr = NULL)
}
\arguments{
\item{gen}{Genome to retrieve, currently only hg38, hg19, mm9, mm10 (default: hg38)}

\item{chr}{Chromosome to retrieve length of, all chromosomes if NULL (default: NULL)}
}
\value{
seqlengths of gen
}
\description{
Get the seqlengths of a chromosome or all chromosomes
}
\examples{

gen.sl <- getSeqLengths("hg38", "chr11")

}
