% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth1DDensity.R
\name{meth1DDensity}
\alias{meth1DDensity}
\title{Create a 1D density plot of methylation level}
\usage{
meth1DDensity(betas, line.color = "red", ylabel = "Sample")
}
\arguments{
\item{betas}{BSseq or matrix of methylation levels (assumes matrix colnames are the sample names)}

\item{line.color}{Color of plotted line (default: "red")}

\item{ylabel}{Y-axis label (default: "Sample")}
}
\value{
a ggplot2 object of the density plot
}
\description{
Create a 1D density plot of methylation level
}
\examples{

orig_bed <- system.file("extdata", "MCF7_Cunha_chr11p15.bed.gz",
                        package="biscuiteer")
orig_vcf <- system.file("extdata", "MCF7_Cunha_header_only.vcf.gz",
                        package="biscuiteer")
bisc <- readBiscuit(BEDfile = orig_bed, VCFfile = orig_vcf,
                    merged = FALSE)
meth1DDensity(bisc)

}
