% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscalePlot.R
\name{multiscaleGroupAverage}
\alias{multiscaleGroupAverage}
\title{Create a list of GRangesLists for each step in the multiscale plot}
\usage{
multiscaleGroupAverage(means_dir, samples, config, which)
}
\arguments{
\item{means_dir}{List of length nSamples. Each element of means.list is itself a list of GRanges of length, nSteps}

\item{samples}{character vector of samples to include. The must match the sample column from samples.tsv in the
multiscale plot workflow}

\item{config}{Config YAML from the multiscale plot workflow}

\item{which}{GRanges for rtracklayer::import}
}
\value{
a GRangesList with the average values
}
\description{
Output of this function can be passed directly to multiscaleMethylationPlot
}
\examples{

# Not a great example, but it works
files.loc <- system.file("extdata", package="bisplotti")

grp_avg <- multiscaleGroupAverage(
    files.loc,
    "Heyn_2012_100yr_subset",
    paste0(files.loc, "/config_example.yaml"),
    NULL
)

}
