% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscalePlot.R
\name{multiscaleMethylationPlot}
\alias{multiscaleMethylationPlot}
\title{Create the multiscale plot from a list of GRanges}
\usage{
multiscaleMethylationPlot(
  grl,
  colors = "YlGnBu",
  na_color = "darkgray",
  what = "score",
  chr.to.plot = NULL
)
}
\arguments{
\item{grl}{GRangesList of a single set of multiscale BED files}

\item{colors}{RColorBrewer color scheme to use (default: YlGnBu)}

\item{na_color}{Color to give NA values (default: darkgray)}

\item{what}{Column to plot (default: score)}

\item{chr.to.plot}{Chromosome to plot, all if NULL (default: NULL)}
}
\value{
a ggplot2 object of the multiscale plot
}
\description{
Create the multiscale plot from a list of GRanges
}
\examples{

}
