% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epistate.R
\name{plotEpiread}
\alias{plotEpiread}
\title{Plot the results of tabulateEpibed() as a
quasi-lollipop plot.}
\usage{
plotEpiread(
  mat,
  plot_read_ave = TRUE,
  show_readnames = TRUE,
  show_positions = TRUE,
  unmeth_color = "white",
  meth_color = "black",
  na_color = "darkgray",
  a_color = "#F26419",
  t_color = "#E9C46A",
  g_color = "#264653",
  c_color = "#2A9D8F",
  background_color = "#A3D0E9"
)
}
\arguments{
\item{mat}{Input matrix that comes out of tabulateEpibed()}

\item{plot_read_ave}{Whether to also plot the average methylation state (default: TRUE)}

\item{show_readnames}{Whether to show the read names (default: TRUE)}

\item{show_positions}{Whether to show the genomic positions (default: TRUE)}

\item{unmeth_color}{What color should the unmethylated states be (default: 'white')}

\item{meth_color}{What color should the methylated states be (default: 'black')}

\item{a_color}{What color should the A SNPs be (default: '#F26419')}

\item{t_color}{What color should the T SNPs be (default: '#E9C46A')}

\item{g_color}{What color should the G SNPs be (default: '#264653')}

\item{c_color}{What color should the C SNPs be (default: '#2A9D8F')}

\item{na_color}{What color should the NA values be (default: 'darkgray')}

\item{background_color}{What color the background of the plot should be (default: '#A3D0E9')}
}
\value{
An epiread ggplot object or list of ggplot objects if plot_read_ave is TRUE
}
\description{
Plot the results of tabulateEpibed() as a
quasi-lollipop plot.
}
\examples{
epibed.nome <- system.file("extdata", "hct116.nome.epiread",
                           package="biscuiteer")
epibed.nome.gr <- readEpibed(epibed = epibed.nome, is.nome = TRUE,
                             genome = "hg19")
epibed.tab.nome <- tabulateEpibed(epibed.nome.gr)
plotEpiread(epibed.tab.nome$gc_table)
}
