% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscalePlot.R
\name{multiscaleMethylationPlot}
\alias{multiscaleMethylationPlot}
\title{Create the multiscale plot from a GRangesList}
\usage{
multiscaleMethylationPlot(
  grl,
  colors = "YlGnBu",
  na_color = "darkgray",
  what = "score",
  config = NULL,
  chr.to.plot = NULL
)
}
\arguments{
\item{grl}{GRangesList of a single set of multiscale BED files}

\item{colors}{RColorBrewer or viridis color scheme to use (default: YlGnBu)}

\item{na_color}{Color to give NA values (default: darkgray)}

\item{what}{Column to plot (default: score)}

\item{config}{path to the bin/config.yaml of the multiscale Snakemake}

\item{chr.to.plot}{Chromosome to plot, all if NULL (default: NULL)}
}
\value{
a ggplot2 object of the multiscale plot
}
\description{
Create the multiscale plot from a GRangesList
}
\examples{

files.loc <- system.file("extdata", package="bisplotti")

files <- lapply(
    list.files(files.loc, pattern="Heyn_2012_100yr", full.names=TRUE), function(x) {
        return(rtracklayer::import(x, format="bedGraph"))
    }
)
cnames <- list.files(files.loc, pattern="Heyn_2012_100yr")
cnames <- gsub("Heyn_2012_100yr_subset_", "", cnames)
cnames <- gsub(".bed.gz", "", cnames)
names(files) <- as.numeric(cnames)

files.grl <- as(files, "GRangesList")

multiscaleMethylationPlot(files.grl)

}
